requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
boto3
requests
snowflake-connector-python
dbt-core
dbt-snowflake
apache-airflow-providers-amazon>=8.0.0
apache-airflow-providers-snowflake
